feat: ✨ update to 19.5.435

+++ Runtime/Asset/HEU_AssetUpdater.cs
@@ -36,6 +36,15 @@ using UnityEngine;
 using UnityEditor;
 #endif
 
+// Expose internal classes/functions
+#if UNITY_EDITOR
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("HoudiniEngineUnityEditor")]
+[assembly: InternalsVisibleTo("HoudiniEngineUnityEditorTests")]
+[assembly: InternalsVisibleTo("HoudiniEngineUnityPlayModeTests")]
+#endif
+
 namespace HoudiniEngineUnity
 {
     /// <summary>
@@ -45,7 +54,7 @@ namespace HoudiniEngineUnity
 #if UNITY_EDITOR && HOUDINIENGINEUNITY_ENABLED
     [InitializeOnLoad]
 #endif
-    public class HEU_AssetUpdater
+    internal class HEU_AssetUpdater
     {
 #if UNITY_EDITOR && HOUDINIENGINEUNITY_ENABLED
 	private static List<HEU_HoudiniAsset> _allHoudiniAssets = new List<HEU_HoudiniAsset>();
@@ -167,7 +176,7 @@ namespace HoudiniEngineUnity
 	/// so this notifies user and provides a way to clean up the created prefab.
 	/// </summary>
 	/// <param name="instance">New prefab instance that was created</param>
-	static void OnPrefabInstanceUpdate(GameObject instance)
+	private static void OnPrefabInstanceUpdate(GameObject instance)
 	{
 #if UNITY_EDITOR && HOUDINIENGINEUNITY_ENABLED && UNITY_2017_1_OR_NEWER
 
@@ -186,11 +195,13 @@ namespace HoudiniEngineUnity
 				"Prefab: " + prefabPath;
 
 		heu_root._houdiniAsset.WarnedPrefabNotSupported = true;
-		if (HEU_EditorUtility.DisplayDialog(title, message, "Remove Prefab & Revert", "Keep Prefab"))
+		if (HEU_EditorUtility.DisplayDialog(title, message, "Remove Prefab && Revert", "Keep Prefab"))
 		{
 		    HEU_EditorUtility.DisconnectPrefabInstance(instance);
-
 		    HEU_AssetDatabase.DeleteAssetAtPath(prefabPath);
+
+		    // Doesn't actually refresh properly, but should be fixed in 2021.2 and 2022.1 
+		    HEU_AssetDatabase.SaveAndRefreshDatabase();
 		}
 	    }
 #endif
